+++ ezdxf/options.py
@@ -10,9 +10,6 @@ class Options(object):
     def __init__(self):
         self.template_dir = None
 
-        # compress tags of unknown chunks of tags like unknown sections
-        self.compress_default_chunks = False
-
         # compress binary data tags 310-319, collects multiple succeeding binary tags as one compressed tag
         self.compress_binary_data = False